moved duplicate code to trait
