+++ core/src/main/scala/com/salesforce/op/stages/impl/classification/MultiClassificationModelSelector.scala
@@ -33,7 +33,7 @@ package com.salesforce.op.stages.impl.classification
 import com.salesforce.op.evaluators._
 import com.salesforce.op.stages.impl.ModelsToTry
 import com.salesforce.op.stages.impl.classification.{MultiClassClassificationModelsToTry => MTT}
-import com.salesforce.op.stages.impl.selector.{DefaultSelectorParams, ModelSelector}
+import com.salesforce.op.stages.impl.selector.{DefaultSelectorParams, MakeSelector, ModelSelector}
 import com.salesforce.op.stages.impl.tuning._
 import enumeratum.Enum
 import com.salesforce.op.stages.impl.selector.ModelSelectorNames.{EstimatorType, ModelType}
@@ -44,12 +44,12 @@ import org.apache.spark.ml.tuning.ParamGridBuilder
 /**
  * A factory for Multi Classification Model Selector
  */
-case object MultiClassificationModelSelector {
+case object MultiClassificationModelSelector extends MakeSelector {
 
   private[op] val modelNames: Seq[MultiClassClassificationModelsToTry] = Seq(MTT.OpLogisticRegression,
     MTT.OpRandomForestClassifier) // OpDecisionTreeClassifier and OpNaiveBayes off by default
 
-  private def defaultModelsAndParams: Seq[(EstimatorType, Array[ParamMap])] = {
+  protected def defaultModelsAndParams: Seq[(EstimatorType, Array[ParamMap])] = {
     val lr = new OpLogisticRegression()
     val lrParams = new ParamGridBuilder()
       .addGrid(lr.fitIntercept, DefaultSelectorParams.FitIntercept)
@@ -174,32 +174,8 @@ case object MultiClassificationModelSelector {
       trainTestEvaluators = Seq(new OpMultiClassificationEvaluator) ++ trainTestEvaluators,
       modelTypesToUse = modelTypesToUse, modelsAndParameters = modelsAndParameters)
   }
-
-  private def selector(
-    validator: OpValidator[ModelType, EstimatorType],
-    splitter: Option[DataCutter],
-    trainTestEvaluators: Seq[OpMultiClassificationEvaluatorBase[_ <: EvaluationMetrics]],
-    modelTypesToUse: Seq[MultiClassClassificationModelsToTry],
-    modelsAndParameters: Seq[(EstimatorType, Array[ParamMap])]
-  ): ModelSelector[ModelType, EstimatorType] = {
-    val modelStrings = modelTypesToUse.map(_.entryName)
-    val modelsToUse =
-      // if no models are specified use the defaults and filter by the named models to use
-      if (modelsAndParameters.isEmpty) defaultModelsAndParams
-        .filter{ case (e, p) => modelStrings.contains(e.getClass.getSimpleName) }
-      // if models to use has been specified and the models have been specified filter the models by the names
-      else if (modelTypesToUse != modelNames) modelsAndParameters
-        .filter{ case (e, p) => modelStrings.contains(e.getClass.getSimpleName) }
-      // else just use the specified models
-      else modelsAndParameters
-    new ModelSelector(
-      validator = validator,
-      splitter = splitter,
-      models = modelsToUse,
-      evaluators = trainTestEvaluators
-    )
-  }
 }
+
 /**
  * Enumeration of possible classification models in Model Selector
  */
